Rename aws-lambda file
