+++ src/handlers/platforms/aws-lambda/create-app-register-handler.ts
