+++ src/handlers/platforms/aws-lambda/index.ts
@@ -1,4 +1,4 @@
-export * from "./create-app-register";
+export * from "./create-app-register-handler";
 export * from "./create-manifest-handler";
 export * from "./create-protected-handler";
 export * from "./platform-adapter";
